refactor(scss): move sass into the source folder
