+++ demos/demo-app.module.ts
@@ -1,7 +1,7 @@
 import 'bootstrap/dist/css/bootstrap.css';
 import 'font-awesome/css/font-awesome.css';
 import 'highlight.js/styles/github.css';
-import '../scss/angular-calendar.scss';
+import '../src/angular-calendar.scss';
 import { NgModule } from '@angular/core';
 import { RouterModule, PreloadAllModules } from '@angular/router';
 import { BrowserModule } from '@angular/platform-browser';
