+++ package.json
@@ -18,7 +18,7 @@
     "commit": "git-cz",
     "compodoc": "compodoc -p tsconfig-compodoc.json -d docs --disableGraph --disableCoverage --disablePrivateOrInternalSupport",
     "gh-pages": "git checkout gh-pages && git merge master --no-edit && rm -rf *.js *.js.map && npm run build:demos && npm run compodoc && git add . && git commit -m 'chore: build demos and docs' && git push && git checkout master",
-    "copyfiles": "copyfiles package.json LICENSE README.md CHANGELOG.md scss/*.scss dist",
+    "copyfiles": "copyfiles package.json LICENSE README.md CHANGELOG.md dist && copyfiles -u 1 src/**/*.scss dist/scss",
     "prerelease": "npm test",
     "release:git": "git add package.json package-lock.json && git commit -m 'chore: bump version number' && standard-version --first-release && git push --follow-tags origin master",
     "release:npm": "npm run build:dist && npm run copyfiles && npm publish dist",
