--- scss/angular-calendar.scss
@@ -1,4 +1,0 @@
-@import 'month-view';
-@import 'week-view';
-@import 'day-view';
-@import 'tooltip';
