+++ src/angular-calendar.scss
@@ -1,0 +1,4 @@
+@import './modules/month/calendar-month-view';
+@import './modules/week/calendar-week-view';
+@import './modules/day/calendar-day-view';
+@import './modules/common/calendar-tooltip';
