+++ src/modules/common/calendar-tooltip.scss
