+++ src/modules/day/calendar-day-view.scss
