+++ src/modules/month/calendar-month-view.scss
