+++ src/modules/week/calendar-week-view.scss
